refactored docs
