+++ docs/index.rst
@@ -1,13 +1,15 @@
 pyansys Documentation
 =====================
-This Python module allows you to:
- - Interactively control an instance of ANSYS v17.0 and newer using
-   Python.
+The ``pyansys`` provides various pythonic interfaces to the MAPDL
+application and both ASCII and binary MAPDL files.  It allows you to:
+
+ - Interactively control an instance of ANSYS v17.0 on and newer using
+   Python on Windows or v13 and newer on Linux.
  - Extract data directly from binary ANSYS v14.5+ files and display or
    animate them.
- - Rapidly read in binary results (``.rst``, ``.rth``), binary mass and stiffness
-   ``(.full)``, and ASCII block archive ``(.cdb)``, and element matrix
-   ``(.emat)`` files.
+ - Rapidly read in binary results (``.rst``, ``.rth``), binary mass
+   and stiffness ``(.full)``, and ASCII block archive ``(.cdb)``, and
+   element matrix ``(.emat)`` files.
 
 This python module is a community driven work in progress with
 additional features regularly added based on user requests.  Open an
@@ -16,40 +18,6 @@ if you identity a bug or would like to request an additional method or
 feature.
 
 Take a look at the :ref:`sphx_glr_examples` gallery for several demos using ``pyansys``.
-
-
-Contents
---------
-
-.. toctree::
-   :maxdepth: 1
-   :caption: Interactive MAPDL Python Interface
-
-   mapdl/index
-   mapdl/mapdl
-
-.. toctree::
-   :maxdepth: 1
-   :caption: ANSYS File Support
-
-   archive
-   loading_results
-   examples
-   loading_km
-   loading_emat
-
-.. toctree::
-   :maxdepth: 1
-   :caption: Example Gallery
-
-   examples/index
-
-.. toctree::
-   :maxdepth: 1
-   :caption: Miscellaneous
-   :hidden:
-
-   quality
 
 
 Installation
@@ -68,7 +36,10 @@ Quick Examples
 
 Interactive Control
 ~~~~~~~~~~~~~~~~~~~
-``pyansys`` lets you create an instance of ANSYS and send commands to it pythonically:
+You can create an interactive instance of MAPDL using ``pyansys`` to
+send commands to it Pythonically, plot meshes and geometry within
+python, and access underlying results directly through python.  Here's
+a sample code snippet.
 
 .. code:: python
 
@@ -126,19 +97,59 @@ plots results directly from an ANSYS result file using ``pyansys``:
 .. figure:: ./images/rotor.jpg
     :width: 500pt
 
+--------
+
+Contents
+========
+.. toctree::
+   :maxdepth: 1
+   :caption: Interactive MAPDL Python Interface
+
+   mapdl/index
+   mapdl/mapdl
+   mapdl/plotting
+   mapdl/parameters
+   mapdl/conversion
+   mapdl/functions
+   mapdl/examples
+
+.. toctree::
+   :maxdepth: 1
+   :caption: ANSYS File Support
+
+   archive
+   loading_results
+   examples
+   loading_km
+   loading_emat
+
+.. toctree::
+   :maxdepth: 1
+   :caption: Example Gallery
+
+   examples/index
+
+.. toctree::
+   :maxdepth: 1
+   :caption: Miscellaneous
+   :hidden:
+
+   quality
+
+--------
 
 License and Acknowledgments
 ---------------------------
 The ``pyansys`` module is licensed under the MIT license.
 
-ANSYS documentation and functions build from html provided by
+ANSYS documentation and functions build from html that was hosted on
 `Sharcnet <https://www.sharcnet.ca/Software/Ansys/>`_.  They've since
-taken down their documentation, but it was a great resource and I
+taken down their documentation, but it was an excellent resource and I
 appreciate their work.
 
 This module, ``pyansys`` makes no commercial claim over ANSYS
-whatsoever.  This tool extends the functionality of ``MAPDL`` by
-adding a python interface in both file interface as well as
-interactive scripting without changing the core behavior or license of
-the original software.  The interactive use of MAPDL using ``pyansys``
-requires a legally licensed local copy of ANSYS.
+whatsoever.  This tool extends the functionality of MAPDL by adding a
+Python interface to both the core application and the binary files
+behavior or license of the original software.  The interactive use of
+MAPDL using ``pyansys`` requires a legally licensed local copy of
+ANSYS.
