+++ docs/mapdl/examples.rst
